[ALPS03794734] MSDC:MT6761: fix coverity issue
[Detail]
Unintentional integer overflow (OVERFLOW_BEFORE_WIDEN)
To avoid overflow, cast either ms or 1000U to type u64.

[Solution]
add type u64 to ms

MTK-Commit-Id: 8a646784b462c2b6c2bd20c15d15dc2d421ee9c2

CR-Id: ALPS03794734
Feature: eMMC Boot Up
(cherry picked from commit df4f90956197c8a722046915e6d8e1356d56117c)

+++ drivers/mmc/host/mediatek/ComboA/sd.c
@@ -727,7 +727,7 @@ static void msdc_set_busy_timeout_ms(struct msdc_host *host, u32 ms)
 	u64 timeout, clk_ns, us;
 	u32 mode = 0;
 
-	us = ms * 1000;
+	us = (u64)ms * 1000;
 
 	if (host->sclk == 0) {
 		timeout = 0;
